Abstract over logging a bit
Each of the loggers were doing the same thing:
someting before processing the request, and something after processing.
We can pull that out to a base `log` middleware,
and make things a bit more extensible.

We provide a helper for time as well.
Since each of the loggers thus far were doing time things,
it seems reasonable to have that information available if you want it.
We don't have to force everyone to know how to time requests.