Lock down `purescript-httpure` version
In SemVer, the semantics of `0.x.x` versions is unspecified. Many
projects decide their own versioning scheme. `purescript-httpure`
decided that breaking changes are fine in patch versions.

This versioning scheme doesn't work with the caret syntax of SemVer, so
we lock down to an exact version. If `purescript-httpure` decides on a
different versioning scheme, we can change this.