Bring all `bower` dependencies up to date
We need these dependencies in order to work with 0.14.0.

The only one that's a bit hard is the `purescript-httpure` dependency.
The actual version should be 0.11.0, but that tag doesn't have the
correct `bower` dependencies. So, we reference a commit that does have
the correct `bower` dependencies. Hopefully we'll get a new tag we can
reference instead:
https://github.com/cprussin/purescript-httpure/pull/168#issuecomment-803415976.